changes for mqx:
	- connector_api_file_system.h: int32_t is defined at stdint.h, not included yet...
	- connector_types.h: don't include stdio.h in IAR, it conflicts with fio.h (MQX only...).
	- connector_api.h: CONNECTOR_VERSION migth be defined somewhere else
	- connector_api.c: init allocated memmory to 0x00.

+++ public/include/connector_api_file_system.h
@@ -63,7 +63,7 @@ typedef enum {
 #if (defined CONNECTOR_FILE_SYSTEM_HAS_LARGE_FILES)
 typedef int64_t connector_file_offset_t;
 #else
-typedef int32_t connector_file_offset_t;
+typedef size_t connector_file_offset_t;
 #endif
 /**
 * @}
